+++ examples/library.rs
@@ -6,11 +6,11 @@
 use anyhow::Result;
 use bliss_audio::decoder::ffmpeg::FFmpegDecoder as Decoder;
 use bliss_audio::library::{AppConfigTrait, BaseConfig, Library};
+use bliss_audio::AnalysisOptions;
 use clap::{value_parser, Arg, Command};
 use glob::glob;
 use serde::{Deserialize, Serialize};
 use std::fs;
-use std::num::NonZeroUsize;
 use std::path::{Path, PathBuf};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
@@ -32,9 +32,9 @@ impl Config {
         music_library_path: PathBuf,
         config_path: Option<PathBuf>,
         database_path: Option<PathBuf>,
-        number_cores: Option<NonZeroUsize>,
+        analysis_options: Option<AnalysisOptions>,
     ) -> Result<Self> {
-        let base_config = BaseConfig::new(config_path, database_path, number_cores)?;
+        let base_config = BaseConfig::new(config_path, database_path, analysis_options)?;
         Ok(Self {
             base_config,
             music_library_path,